feat: setup icon
