+++ Sources/App/Main.swift
@@ -1,0 +1,16 @@
+import AppKit
+
+@main
+enum Main {
+    static func main() {
+        if CommandLine.arguments.contains("--selftest") {
+            SelfTest.run()
+            return
+        }
+        let app = NSApplication.shared
+        let delegate = AppDelegate()
+        app.delegate = delegate
+        app.setActivationPolicy(.accessory)  // menu-bar agent, no dock icon
+        app.run()
+    }
+}
